+++ index.d.ts
@@ -1,15 +1,23 @@
 import { IInit } from './src/model';
 import { IInstabug } from './src/model/instabug';
 import { ISentry } from './src/model/sentry';
+import {ITealium} from "./src/model/tealium";
+
+declare const DEBUG_LOG: number;
+declare const WARNING_LOG: number;
+declare const ERROR_LOG: number;
+declare const NETWORK_LOG: number;
 
 declare const init: (initConfig: IInit) => void;
 
-declare const setupReactotron: (config?: any, plugins?: Array<Function>) => void;
+declare const setupReactotron: (config?: any, plugins?: Array<Function>) => any;
 
-declare const createFirebaseLogger: (analytics: any, printError?: boolean) => void;
-declare const createCrashlyticsLogger: (crashlytics: any, printError?: boolean) => void;
-declare const createSentryLogger: (sentry: any, config: ISentry, printError?: boolean) => void;
-declare const createInstabugLogger: (instabug: any, config: IInstabug, printError?: boolean) => void;
+declare const createTealiumLogger: (Tealium: any, config: ITealium, printLogs?: boolean) => Function;
+declare const createAdobeLogger: (ACPAnalytics: any, ACPCore: any, printLogs?: boolean) => Function;
+declare const createFirebaseLogger: (analytics: any, printLogs?: boolean) => Function;
+declare const createCrashlyticsLogger: (crashlytics: any, printLogs?: boolean) => Function;
+declare const createSentryLogger: (sentry: any, config: ISentry, printLogs?: boolean) => Function;
+declare const createInstabugLogger: (instabug: any, config: IInstabug, printLogs?: boolean) => Function;
 
 declare const logEvent: (event: string, params?: any, sensitiveData?: boolean) => void;
 declare const logDebugEvent: (event: string, params?: any, sensitiveData?: boolean) => void;