+++ src/__tests__/index.spec.ts
@@ -2,6 +2,8 @@ import {
   init,
   createSentryLogger,
   createFirebaseLogger,
+  createAdobeLogger,
+  createTealiumLogger,
   setupReactotron,
   logEvent,
   createCrashlyticsLogger,
@@ -11,13 +13,23 @@ import {
   logNetworkEvent,
   logErrorEvent,
   createInstabugLogger,
+  DEBUG_LOG,
 } from '../index';
 import * as Init from '../modules/init';
-import { emptyFunction } from '../modules/reactotron';
+import { ITealium } from '../model/tealium';
+import { log } from '../modules/events';
+import { excludeLogs, loggers } from '../modules/init';
 
 describe('index test suite', () => {
   const analytics = {
     logEvent: jest.fn(),
+  };
+  const ACPAnalytics = {
+    registerExtension: jest.fn(),
+  };
+  const ACPCore = {
+    trackState: jest.fn(),
+    trackAction: jest.fn(),
   };
   const instabug = {
     startWithToken: jest.fn(),
@@ -40,6 +52,12 @@ describe('index test suite', () => {
     init: jest.fn(),
     captureMessage: jest.fn(),
   };
+  const Tealium = {
+    initialize: jest.fn(),
+    trackEvent: jest.fn(),
+  };
+  const configTealium: ITealium = { account: 'accountName', profile: 'profileName', environment: 'environment' };
+
   const AsyncStorage = jest.fn();
   const reactotronRedux = jest.fn();
   const Reactotron = {
@@ -48,24 +66,63 @@ describe('index test suite', () => {
     useReactNative: () => Reactotron,
     use: () => Reactotron,
     connect: () => Reactotron,
-    clear: () => Reactotron,
     createEnhancer: () => Reactotron,
   };
 
-  it('should init properly', () => {
-    init({});
+  beforeEach(() => {
+    jest.resetAllMocks();
   });
 
-  it('should init not properly', () => {
+  it('should init properly empty', () => {
+    init({});
+    expect(loggers).toEqual([]);
+  });
+
+  it('should init properly', () => {
     init({ config: {}, analytics: [], errorReporters: [] });
+    expect(loggers).toEqual([]);
+  });
+
+  it('should init properly with wrong analytics and errorReporters', () => {
+    // @ts-ignore
+    init({ config: {}, analytics: [{}], errorReporters: [{}] });
+    expect(loggers).toEqual([]);
+  });
+
+  it('should init properly and log event with reject', () => {
+    init({
+      config: {
+        excludeLogs: {
+          instabug: [DEBUG_LOG],
+          tealium: [DEBUG_LOG],
+          adobe: [DEBUG_LOG],
+          firebase: [DEBUG_LOG],
+          sentry: [DEBUG_LOG],
+          crashlytics: [DEBUG_LOG],
+        },
+      },
+      analytics: [
+        createInstabugLogger(instabug, { invocationEvent: 'shake', token: 'token' }),
+        createTealiumLogger(Tealium, configTealium),
+        createAdobeLogger(ACPAnalytics, ACPCore),
+        createFirebaseLogger({}),
+        createSentryLogger({ init: jest.fn() }, { dsn: 'dsn' }),
+        createCrashlyticsLogger(crashlytics),
+      ],
+    });
+    logDebugEvent('event');
   });
 
   it('should init not properly 2', () => {
-    // @ts-ignore
-    init({ config: { reportJSErrors: true, isSensitiveBuild: true }, analytics: [{}], errorReporters: [{}] });
+    init({
+      config: { reportJSErrors: true, isSensitiveBuild: true, excludeLogs: { instabug: [DEBUG_LOG] } },
+    });
+    expect(excludeLogs).toEqual({ instabug: [DEBUG_LOG] });
   });
 
-  it('should init with wrong datas', () => {
+  // Firebase & Sentry
+
+  it('should init Firebase & Sentry with wrong datas + printLogs', () => {
     init({
       analytics: [createFirebaseLogger({}, true), createSentryLogger({ init: jest.fn() }, { dsn: 'dsn' }, true)],
       errorReporters: [createCrashlyticsLogger({}, true)],
@@ -74,7 +131,7 @@ describe('index test suite', () => {
     recordError('error', { key: 'value' });
   });
 
-  it('should init with wrong datas', () => {
+  it('should init Firebase & Sentry with wrong datas', () => {
     init({
       analytics: [createFirebaseLogger({}), createSentryLogger({ init: jest.fn() }, { dsn: 'dsn' })],
       errorReporters: [createCrashlyticsLogger({})],
@@ -82,8 +139,6 @@ describe('index test suite', () => {
     logEvent('event', { key: 'value' });
     recordError('error', { key: 'value' });
   });
-
-  // Firebase & Sentry
 
   it('should init properly and log event', () => {
     givenSetSensitiveBuild(true);
@@ -105,30 +160,70 @@ describe('index test suite', () => {
     logErrorEvent('event');
   });
 
-  // Instabug
+  // Tealium & Adobe
 
-  it('should init properly and log event', () => {
+  it('should init Tealium & Adobe with wrong datas + printLogs', () => {
     init({
-      analytics: [createInstabugLogger(instabug, { invocationEvent: 'shake', token: 'token' })],
+      analytics: [
+        createTealiumLogger({ initialize: jest.fn() }, configTealium, true),
+        createAdobeLogger({ registerExtension: jest.fn() }, {}, true),
+      ],
+      errorReporters: [createCrashlyticsLogger({}, true)],
     });
-    logEvent('event', { key: 'value' }, true);
+    logEvent('event', { key: 'value' });
+    recordError('error', { key: 'value' });
   });
 
-  it('should init properly and log event without invocationEvent', () => {
+  it('should init Tealium & Adobe with wrong datas', () => {
+    init({
+      analytics: [
+        createTealiumLogger({ initialize: jest.fn() }, configTealium),
+        createAdobeLogger({ registerExtension: jest.fn() }, {}),
+      ],
+      errorReporters: [createCrashlyticsLogger({})],
+    });
+    logEvent('event', { key: 'value' });
+    recordError('error', { key: 'value' });
+  });
+
+  // Firebase & Sentry & Adobe & Tealium
+
+  it('should init properly and log event', () => {
+    givenSetSensitiveBuild(true);
+    init({
+      analytics: [createTealiumLogger(Tealium, configTealium), createAdobeLogger(ACPAnalytics, ACPCore)],
+    });
+    log('event', -1);
+  });
+
+  it('should init properly and log without param', () => {
+    givenSetSensitiveBuild(false);
+    init({
+      analytics: [createTealiumLogger(Tealium, configTealium), createAdobeLogger(ACPAnalytics, ACPCore)],
+    });
+    logEvent('event');
+    logWarningEvent('event');
+    logDebugEvent('event');
+    logNetworkEvent('event');
+    logErrorEvent('event');
+  });
+
+  // Instabug
+  it('should init Instabug properly and log event without invocationEvent', () => {
     init({
       analytics: [createInstabugLogger(instabug, { token: 'token' })],
     });
     logEvent('event', { key: 'value' }, true);
   });
 
-  it('should init not properly', () => {
+  it('should init Instabug not properly + printLogs', () => {
     init({
       analytics: [createInstabugLogger(instabugWrong, { token: 'token' }, true)],
     });
     logEvent('event', { key: 'value' }, true);
   });
 
-  it('should init not properly', () => {
+  it('should init Instabug not properly', () => {
     init({
       analytics: [createInstabugLogger(instabugWrong, { token: 'token' })],
     });
@@ -145,18 +240,6 @@ describe('index test suite', () => {
   it('should init properly and record error without param', () => {
     init({ config: { reportJSErrors: true }, errorReporters: [createCrashlyticsLogger(crashlytics)] });
     recordError('error');
-  });
-
-  it('should init properly and setup reactotron', () => {
-    init({ config: { Reactotron, AsyncStorage } });
-    emptyFunction();
-    setupReactotron();
-  });
-
-  it('should init properly and setup reactotron and redux', () => {
-    init({ config: { Reactotron, AsyncStorage } });
-    setupReactotron(undefined, [reactotronRedux()]);
-    setupReactotron({}, [reactotronRedux(), reactotronRedux(), reactotronRedux()]);
   });
 
   function givenSetSensitiveBuild(isSensitive: boolean = true) {