+++ src/index.ts
@@ -8,18 +8,22 @@ import {
   recordError,
 } from './modules/events';
 import {
+  createTealiumLogger,
+  createAdobeLogger,
   createFirebaseLogger,
   createCrashlyticsLogger,
   createSentryLogger,
   createInstabugLogger,
 } from './modules/loggers';
-import { init } from './modules/init';
+import { init, DEBUG_LOG, WARNING_LOG, ERROR_LOG, NETWORK_LOG } from './modules/init';
 
 export {
   createFirebaseLogger,
   createSentryLogger,
   createCrashlyticsLogger,
   createInstabugLogger,
+  createTealiumLogger,
+  createAdobeLogger,
   setupReactotron,
   init,
   logEvent,
@@ -28,4 +32,8 @@ export {
   logNetworkEvent,
   logWarningEvent,
   recordError,
+  DEBUG_LOG,
+  WARNING_LOG,
+  ERROR_LOG,
+  NETWORK_LOG,
 };