+++ src/modules/loggers/index.ts
@@ -2,5 +2,14 @@ import { createFirebaseLogger } from './firebase';
 import { createSentryLogger } from './sentry';
 import { createCrashlyticsLogger } from './crashlytics';
 import { createInstabugLogger } from './instabug';
+import { createAdobeLogger } from './adobe';
+import { createTealiumLogger } from './tealium';
 
-export { createFirebaseLogger, createSentryLogger, createCrashlyticsLogger, createInstabugLogger };
+export {
+  createFirebaseLogger,
+  createSentryLogger,
+  createCrashlyticsLogger,
+  createInstabugLogger,
+  createAdobeLogger,
+  createTealiumLogger,
+};